+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include README.md
 include LICENSE
+include CHANGELOG.md
