+++ src/mongodesu/__init__.py
@@ -1,0 +1,3 @@
+from .mongolib import MongoAPI, Model
+
+__all__ = ["MongoAPI", "Model"]
