+++ src/mongodesu/serializable/__init__.py
@@ -1,0 +1,3 @@
+from .serializable import Serializable
+
+__all__ = ["Serializable"]
